refractor scala example to latest raphtory update lotr to use latest core change nft folder to match examples update python nft example
