+++ examples/nft/project/plugins.sbt
