+++ examples/nft/python/cyclemania.py
