+++ examples/nft/src/main/scala/com/raphtory/examples/nft/analysis/CycleMania.scala
